requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==39.0.2
Django==4.1.7
pycparser==2.21
pyOpenSSL==23.0.0
sqlparse==0.4.3
